Build: Get building on CentOS 7 w/ correct gcc toolchain (#13)
Issue: reason-skia would fail to link properly on CentOS 7 when built with the OCaml toolchain. There were mismatches in the symbols for thread libraries.

Defect: clang, by default, was pulling in a newer version of the gcc compiler + libraries. This was mismatched with gcc 4.8.5, which we were building with in the ocaml space.

Fix: Match up gcc versions for clang by specifying the -gcc-toolchain option.